+++ runtime/crab/Cargo.toml
@@ -37,7 +37,7 @@ cumulus-pallet-session-benchmarking = { workspace = true, optional = true }
 darwinia-account-migration        = { workspace = true }
 darwinia-common-runtime           = { workspace = true }
 darwinia-deposit                  = { workspace = true }
-darwinia-message-transact         = { workspace = true }
+darwinia-runtime-transact         = { workspace = true }
 darwinia-precompile-assets        = { workspace = true }
 darwinia-precompile-bls12-381     = { workspace = true }
 darwinia-precompile-deposit       = { workspace = true }
@@ -150,7 +150,7 @@ std = [
 	"darwinia-account-migration/std",
 	"darwinia-common-runtime/std",
 	"darwinia-deposit/std",
-	"darwinia-message-transact/std",
+	"darwinia-runtime-transact/std",
 	"darwinia-precompile-assets/std",
 	"darwinia-precompile-bls12-381/std",
 	"darwinia-precompile-deposit/std",
@@ -261,7 +261,7 @@ runtime-benchmarks = [
 	"darwinia-account-migration/runtime-benchmarks",
 	"darwinia-common-runtime/runtime-benchmarks",
 	"darwinia-deposit/runtime-benchmarks",
-	"darwinia-message-transact/runtime-benchmarks",
+	"darwinia-runtime-transact/runtime-benchmarks",
 	"darwinia-precompile-assets/runtime-benchmarks",
 	"darwinia-precompile-staking/runtime-benchmarks",
 	"darwinia-staking/runtime-benchmarks",
@@ -318,7 +318,7 @@ try-runtime = [
 	# darwinia
 	"darwinia-account-migration/try-runtime",
 	"darwinia-deposit/try-runtime",
-	"darwinia-message-transact/try-runtime",
+	"darwinia-runtime-transact/try-runtime",
 	"darwinia-staking/try-runtime",
 
 	# frontier
