+++ runtime/crab/src/pallets/evm.rs
@@ -200,7 +200,7 @@ impl DispatchValidateT<AccountId, RuntimeCall> for DarwiniaDispatchValidator {
 				| RuntimeCall::Vesting(..)
 				| RuntimeCall::Ethereum(..)
 				| RuntimeCall::EVM(..)
-				| RuntimeCall::MessageTransact(..)
+				| RuntimeCall::RuntimeTransact(..)
 		) {
 			Some(fp_evm::PrecompileFailure::Error {
 				exit_status: ExitError::Other(
