+++ runtime/crab/src/pallets/runtime_transact.rs
@@ -19,7 +19,7 @@
 // darwinia
 use crate::*;
 
-impl darwinia_message_transact::Config for Runtime {
-	type LcmpEthOrigin = darwinia_message_transact::EnsureLcmpEthOrigin;
+impl darwinia_runtime_transact::Config for Runtime {
+	type RuntimeEthOrigin = darwinia_runtime_transact::RuntimeRuntimeEthOrigin;
 	type ValidatedTransaction = pallet_ethereum::ValidatedTransaction<Self>;
 }
