+++ runtime/pangolin/src/pallets.rs
@@ -105,4 +105,4 @@ mod ethereum;
 mod evm;
 pub use evm::*;
 
-mod message_transact;
+mod runtime_transact;
