+++ runtime/pangolin/src/pallets/evm.rs
@@ -218,7 +218,7 @@ impl DispatchValidateT<AccountId, RuntimeCall> for DarwiniaDispatchValidator {
 			RuntimeCall::Assets(..)
 				| RuntimeCall::Ethereum(..)
 				| RuntimeCall::EVM(..)
-				| RuntimeCall::MessageTransact(..)
+				| RuntimeCall::RuntimeTransact(..)
 		) {
 			Some(fp_evm::PrecompileFailure::Error {
 				exit_status: ExitError::Other(
