+++ runtime/pangoro/src/pallets.rs
@@ -99,4 +99,4 @@ mod ethereum;
 mod evm;
 pub use evm::*;
 
-mod message_transact;
+mod runtime_transact;
